Clean up the decompiled code in the classes touched by the last two commits
Cosmetic and structural only, no behaviour change except where noted. Scoped
to the 13 classes the rollback/storage work already modified; the rest of the
decompiled tree is left alone rather than half-converted.

Removed decompiler artifacts:

- ~90 redundant casts on Config's already-generic maps. (Integer)Config.config
  .get(k) and friends were casting a Map<String,Integer> to Integer. Checked
  that none of these sat on both sides of an ==, which would have turned value
  comparison into reference comparison; every remaining == compares against an
  int literal and still unboxes.
- Raw (List)/(Map)/(ArrayList) casts on typed fields, and (Material)null /
  (String)null / (List)null argument casts where the callee has no overloads
  to disambiguate.
- Float literals dressed as doubles: (double)0.0F -> 0.0D, (double)0.5F ->
  0.5D, (double)12.5F -> 12.5D. All of these are exactly representable, so the
  values are identical.
- for(/if(/while(/switch(/catch( spacing, and a generated `var9` catch name.
- Kept StringBuilder.length() == 0 rather than isEmpty(); the latter is Java 15
  and this targets 8. String locals did move to isEmpty().

Structural:

- Process.processConsumer was a labelled block with a stray bare scope left
  over from removing the is_paused flip, and 12 prepared statements closed in
  a straight line so a throw part way through leaked the rest. Split into
  processConsumer / resolveUserIds / openBatchStatements / closeQuietly, with
  the statements closed in a finally. Same order of operations, same
  transaction boundaries, same per-row guard -- rows whose user or object is
  missing are skipped and not removed, exactly as before.
- LookupCommand built « / » pagination arrows in three branches whose results
  were never assigned to anything, then overwrote `arrows` with "" before
  sending the header. Dead on arrival; removed. The live total_pages
  calculation further down is untouched.
- UndoCommand, ApplyCommand and CancelCommand were near-identical nests of
  inverted conditionals. Rewritten as guard clauses; the /co apply and
  /co cancel marker rewriting is unchanged.

Left alone deliberately: 3-space indent and cuddled braces, which are the
house style across the whole tree; and the (Integer) casts in CommandHandler,
CoreProtect, CoreProtectAPI and the patch scripts, which this work never
touched.

+++ src/main/java/net/coreprotect/command/PurgeCommand.java
@@ -31,14 +31,14 @@ public class PurgeCommand extends Consumer {
          player.sendMessage(Language.get("please-use-co-purge-t-time"));
       } else if (seconds <= 0) {
          player.sendMessage(Language.get("please-use-co-purge-t-time"));
-      } else if (player instanceof Player && seconds < (Integer)Config.config.get("purge-minimum-time")) {
-         player.sendMessage(Language.get("purge-minimum-age", Functions.formatDuration((Integer)Config.config.get("purge-minimum-time"))));
-      } else if (seconds < (Integer)Config.config.get("purge-minimum-time-console")) {
-         player.sendMessage(Language.get("purge-minimum-age-console", Functions.formatDuration((Integer)Config.config.get("purge-minimum-time-console"))));
+      } else if (player instanceof Player && seconds < Config.config.get("purge-minimum-time")) {
+         player.sendMessage(Language.get("purge-minimum-age", Functions.formatDuration(Config.config.get("purge-minimum-time"))));
+      } else if (seconds < Config.config.get("purge-minimum-time-console")) {
+         player.sendMessage(Language.get("purge-minimum-age-console", Functions.formatDuration(Config.config.get("purge-minimum-time-console"))));
       } else {
          boolean optimizeCheckValue = false;
 
-         for(String arg : args) {
+         for (String arg : args) {
             if (arg.trim().equalsIgnoreCase("#optimize")) {
                optimizeCheckValue = true;
                break;
@@ -55,7 +55,7 @@ public class PurgeCommand extends Consumer {
                   long removed = 0L;
                   Connection connection = null;
 
-                  for(int i = 0; i <= 5; ++i) {
+                  for (int i = 0; i <= 5; ++i) {
                      connection = Database.getConnection(false);
                      if (connection != null) {
                         break;
@@ -73,7 +73,7 @@ public class PurgeCommand extends Consumer {
                   Functions.messageOwnerAndUser(player, Language.get("do-not-restart-your-server-until"));
                   Config.purge_running = true;
 
-                  while(!PurgeCommand.pause_success) {
+                  while (!PurgeCommand.pause_success) {
                      Thread.sleep(1L);
                   }
 
@@ -82,7 +82,7 @@ public class PurgeCommand extends Consumer {
                   PreparedStatement preparedStmt = null;
                   boolean abort = false;
                   String purge_prefix = "tmp_" + Config.prefix;
-                  if ((Integer)Config.config.get("use-mysql") == 0) {
+                  if (Config.config.get("use-mysql") == 0) {
                      query = "ATTACH DATABASE '" + Config.sqlite + ".tmp' AS tmp_db";
                      preparedStmt = connection.prepareStatement(query);
                      preparedStmt.execute();
@@ -99,8 +99,8 @@ public class PurgeCommand extends Consumer {
                      return;
                   }
 
-                  if ((Integer)Config.config.get("use-mysql") == 0) {
-                     for(String table : Config.databaseTables) {
+                  if (Config.config.get("use-mysql") == 0) {
+                     for (String table : Config.databaseTables) {
                         try {
                            query = "DROP TABLE IF EXISTS " + purge_prefix + table + "";
                            preparedStmt = connection.prepareStatement(query);
@@ -116,16 +116,16 @@ public class PurgeCommand extends Consumer {
 
                   List<String> purge_tables = Arrays.asList("sign", "container", "skull", "session", "chat", "command", "entity", "block");
 
-                  for(String table : Config.databaseTables) {
+                  for (String table : Config.databaseTables) {
                      String tableName = table.replace("_", " ");
                      Functions.messageOwnerAndUser(player, Language.get("processing-data", tableName));
-                     if ((Integer)Config.config.get("use-mysql") == 0) {
+                     if (Config.config.get("use-mysql") == 0) {
                         StringBuilder columns = new StringBuilder();
                         ResultSet rs = connection.createStatement().executeQuery("SELECT * FROM " + purge_prefix + table);
                         ResultSetMetaData resultSetMetaData = rs.getMetaData();
                         int columnCount = resultSetMetaData.getColumnCount();
 
-                        for(int i = 1; i <= columnCount; ++i) {
+                        for (int i = 1; i <= columnCount; ++i) {
                            String name = resultSetMetaData.getColumnName(i);
                            if (columns.length() == 0) {
                               columns = new StringBuilder(name);
@@ -205,7 +205,7 @@ public class PurgeCommand extends Consumer {
                            preparedStmt = connection.prepareStatement(query);
 
                            ResultSet resultSet;
-                           for(resultSet = preparedStmt.executeQuery(); resultSet.next(); old_count = resultSet.getInt("count")) {
+                           for (resultSet = preparedStmt.executeQuery(); resultSet.next(); old_count = resultSet.getInt("count")) {
                            }
 
                            resultSet.close();
@@ -221,7 +221,7 @@ public class PurgeCommand extends Consumer {
                            preparedStmt = connection.prepareStatement(query);
 
                            ResultSet resultSet;
-                           for(resultSet = preparedStmt.executeQuery(); resultSet.next(); new_count = resultSet.getInt("count")) {
+                           for (resultSet = preparedStmt.executeQuery(); resultSet.next(); new_count = resultSet.getInt("count")) {
                            }
 
                            resultSet.close();
@@ -233,7 +233,7 @@ public class PurgeCommand extends Consumer {
                         removed += (long)(old_count - new_count);
                      }
 
-                     if ((Integer)Config.config.get("use-mysql") == 1) {
+                     if (Config.config.get("use-mysql") == 1) {
                         try {
                            if (purge_tables.contains(table)) {
                               query = "DELETE FROM " + Config.prefix + table + " WHERE time < '" + ptime + "'";
@@ -248,10 +248,10 @@ public class PurgeCommand extends Consumer {
                      }
                   }
 
-                  if ((Integer)Config.config.get("use-mysql") == 1 && optimizeCheck) {
+                  if (Config.config.get("use-mysql") == 1 && optimizeCheck) {
                      Functions.messageOwnerAndUser(player, Language.get("optimizing-database-please-wait"));
 
-                     for(String table : Config.databaseTables) {
+                     for (String table : Config.databaseTables) {
                         query = "OPTIMIZE LOCAL TABLE " + Config.prefix + table + "";
                         preparedStmt = connection.prepareStatement(query);
                         preparedStmt.execute();
@@ -261,7 +261,7 @@ public class PurgeCommand extends Consumer {
 
                   connection.close();
                   if (abort) {
-                     if ((Integer)Config.config.get("use-mysql") == 0) {
+                     if (Config.config.get("use-mysql") == 0) {
                         (new File(Config.sqlite + ".tmp")).delete();
                         (new File(Config.sqlite + ".tmp-wal")).delete();
                         (new File(Config.sqlite + ".tmp-shm")).delete();
@@ -275,7 +275,7 @@ public class PurgeCommand extends Consumer {
                      return;
                   }
 
-                  if ((Integer)Config.config.get("use-mysql") == 0) {
+                  if (Config.config.get("use-mysql") == 0) {
                      (new File(Config.sqlite)).delete();
                      // In WAL mode the database has -wal and -shm sidecars. They
                      // belong to the file being replaced, and leaving them next
